Implement functionality behind Create and Start buttons
The main GUI has 4 buttons: "Create agents", "Start agents", "Stop agents" and "Exit". "Create agents" is used to load the deployment and create agents within a specific node (the loggingArea is populated with the specific log text), whereas "Start agents" calls start() method on every entity within that node (the logging area is updated and the list of agents from the right of the board is populated with the name of current agents).