Fixed loading of alternate composite agent implementations.
Created the CompositeAgentModel as an interface to any CompositeAgent.

Changed CompositeAgentLoader behavior in case a classpath is present:
- instead of deferring to another loader (which can be covered by
correctly manipulating loader information in the configuration)
- it uses the classpath as an alternative implementation of a
CompositeAgent

Changed the map storing loaders in NodeLoader to a LinkedHashMap, to
keep the order in which loaders have been added (loaders are tried in
order).

Moved SimpleLoader to its own file so that it can be automatically found
using -loader <category>:simple.