NUmerous fixes in how the implementation reflects the principles of the Entity-Operation model.
The main change is that entities no longer have access to FMas, and the
connectTools method has been added to connect them to their
corresponding EntityTools instance.

Created EntityCore that manages some of the basic entity elements (ID,
running state, connection to EntityTools).

Support infrastructure Pylons are now recognized by the fact that they
offer the Route operation, and some may also offer the Register
operation. The request to rout an operation call is now also done by
calling the route operation of the pylon, embedding the original call as
an argument.

Removed the idea of entity name, and left only the ID.

Removed most of the functionality in Node, as it should be covered
elsewhere.

Deprecated the "is routed" flag and associated mechanism for Waves.