Fixed an issue in DeploymentConfiguration.
The issue appeared in the corner case in which a non-default entity
needed to be ported (as part of another, portable, entity). An exception
was being thrown because MultiTreeMap.getTrees only works on multi-keys.

Additionally, improved printing of MultiTreeMap and set a list of used
symbols (strings) at the start of MultiTreeMap.toString.